Convert api_keys to new custom resource format
